🚚 Move notion types to associated files
